refactor(#zimic): improve path transform utilities
