+++ packages/zimic/src/cli/typegen/transform/paths.ts
@@ -13,6 +13,22 @@ export function isPathsDeclaration(node: ts.Node | undefined): node is ts.Interf
   return node !== undefined && ts.isInterfaceDeclaration(node) && node.name.text === 'paths';
 }
 
+function normalizePathNameWithParameters(pathName: string) {
+  return pathName.replace(/{([^}]+)}/g, ':$1');
+}
+
+function wrapComponentPathType(type: ts.TypeNode, context: TypeTransformContext) {
+  context.typeImports.root.add('HttpSchema');
+
+  const httpSchemaMethodsWrapper = ts.factory.createQualifiedName(
+    ts.factory.createIdentifier('HttpSchema'),
+    ts.factory.createIdentifier('Methods'),
+  );
+
+  const wrappedType = ts.factory.createTypeReferenceNode(httpSchemaMethodsWrapper, [type]);
+  return wrappedType;
+}
+
 export function normalizePath(
   path: ts.TypeElement,
   context: TypeTransformContext,
@@ -20,63 +36,49 @@ export function normalizePath(
 ) {
   const { isComponent = false } = options;
 
-  if (ts.isPropertySignature(path) && path.type && ts.isTypeLiteralNode(path.type)) {
-    if (!ts.isStringLiteral(path.name)) {
-      return undefined;
-    }
-
-    const pathName = path.name.text;
-    const newPathName = pathName.replace(/{([^}]+)}/g, ':$1');
-    const newIdentifier = isComponent ? path.name : ts.factory.createStringLiteral(newPathName);
-
-    const newTypeMembers = path.type.members
-      .map((pathMethod) => normalizeMethod(pathMethod, context, { pathName: newPathName }))
-      .filter(isDefined);
-
-    if (newTypeMembers.length === 0) {
-      return undefined;
-    }
-
-    const newType = ts.factory.updateTypeLiteralNode(path.type, ts.factory.createNodeArray(newTypeMembers));
-
-    if (isComponent) {
-      context.typeImports.root.add('HttpSchema');
-
-      const wrappedNewType = ts.factory.createTypeReferenceNode(
-        ts.factory.createQualifiedName(
-          ts.factory.createIdentifier('HttpSchema'),
-          ts.factory.createIdentifier('Methods'),
-        ),
-        [newType],
-      );
-
-      return ts.factory.updatePropertySignature(
-        path,
-        path.modifiers,
-        newIdentifier,
-        path.questionToken,
-        wrappedNewType,
-      );
-    }
-
-    return ts.factory.updatePropertySignature(path, path.modifiers, newIdentifier, path.questionToken, newType);
+  if (
+    !ts.isPropertySignature(path) ||
+    path.type === undefined ||
+    !ts.isTypeLiteralNode(path.type) ||
+    !ts.isStringLiteral(path.name)
+  ) {
+    return undefined;
   }
 
-  return path;
+  const newPathName = isComponent ? path.name.text : normalizePathNameWithParameters(path.name.text);
+  const newIdentifier = isComponent ? path.name : ts.factory.createStringLiteral(newPathName);
+
+  const newTypeMembers = path.type.members
+    .map((pathMethod) => normalizeMethod(pathMethod, context, { pathName: newPathName }))
+    .filter(isDefined);
+
+  if (newTypeMembers.length === 0) {
+    return undefined;
+  }
+
+  let newType: ts.TypeNode = ts.factory.updateTypeLiteralNode(path.type, ts.factory.createNodeArray(newTypeMembers));
+
+  if (isComponent) {
+    newType = wrapComponentPathType(newType, context);
+  }
+
+  const newPath = ts.factory.updatePropertySignature(path, path.modifiers, newIdentifier, path.questionToken, newType);
+  return newPath;
 }
 
 export function normalizePaths(paths: ts.InterfaceDeclaration, context: TypeTransformContext) {
   const newIdentifier = createPathsIdentifier(context.serviceName);
 
   const newMembers = paths.members.map((path) => normalizePath(path, context)).filter(isDefined);
-  const newType = ts.factory.createTypeLiteralNode(ts.factory.createNodeArray(newMembers));
+  const newType = ts.factory.createTypeLiteralNode(newMembers);
 
   context.typeImports.root.add('HttpSchema');
 
-  const wrappedNewType = ts.factory.createTypeReferenceNode(
-    ts.factory.createQualifiedName(ts.factory.createIdentifier('HttpSchema'), ts.factory.createIdentifier('Paths')),
-    [newType],
+  const httpSchemaPathsWrapper = ts.factory.createQualifiedName(
+    ts.factory.createIdentifier('HttpSchema'),
+    ts.factory.createIdentifier('Paths'),
   );
+  const wrappedNewType = ts.factory.createTypeReferenceNode(httpSchemaPathsWrapper, [newType]);
 
   return ts.factory.createTypeAliasDeclaration(paths.modifiers, newIdentifier, paths.typeParameters, wrappedNewType);
 }
