+++ packages/zimic/src/cli/typegen/utils/types.ts
@@ -1,15 +1,35 @@
 import ts from 'typescript';
 
+export function isNeverType(type: ts.TypeNode | undefined): type is undefined {
+  return !type || type.kind === ts.SyntaxKind.NeverKeyword;
+}
+
+export function isNeverTypeMember(member: ts.TypeElement) {
+  return ts.isPropertySignature(member) && isNeverType(member.type);
+}
+
+export function isUnknownType(type: ts.TypeNode) {
+  return type.kind === ts.SyntaxKind.UnknownKeyword;
+}
+
+export function isNumericType(type: ts.TypeNode) {
+  return type.kind === ts.SyntaxKind.NumberKeyword;
+}
+
+export function isBooleanType(type: ts.TypeNode) {
+  return type.kind === ts.SyntaxKind.BooleanKeyword;
+}
+
+export function isNullType(type: ts.TypeNode | ts.LiteralTypeNode['literal']) {
+  return type.kind === ts.SyntaxKind.NullKeyword;
+}
+
 export function createBlobType() {
   return ts.factory.createTypeReferenceNode('Blob');
 }
 
 export function createNullType() {
   return ts.factory.createLiteralTypeNode(ts.factory.createNull());
-}
-
-export function createUnionType(types: ts.TypeNode[]) {
-  return ts.factory.createUnionTypeNode(types);
 }
 
 export function createImportSpecifier(importName: string) {
